fix: Fixing the dynamodb query
Using the timestamp index and first was a bad idea because it return the first lock created in the pool, and if it's not the good lockname, it will return a None. Then, to avoid this behavior, as we filter by lock_name which is unique, we return all results and the filter will do the job.